+++ qancode/defaults.py
@@ -1,0 +1,519 @@
+from .clickpaths import *
+
+# Default browsers.
+BROWSERS = ['Chrome',
+            'Firefox',
+            'Safari']
+
+# Default users.
+USERS = ['Public',
+         'encoded.test@gmail.com',
+         'encoded.test2@gmail.com',
+         'encoded.test4@gmail.com']
+
+
+class bcolors:
+    """
+    Helper class for text color definitions.
+    """
+    OKBLUE = '\x1b[36m'
+    OKGREEN = '\x1b[1;32m'
+    WARNING = '\x1b[33m'
+    FAIL = '\x1b[31m'
+    ENDC = '\x1b[0m'
+
+
+class ActionTuples:
+    """
+    Stores default action lists as attributes.
+    Uses form .{method_name}_default_actions. 
+    """
+
+    def __init__(self):
+        self._init_check_requests_users()
+        self._init_default_actions()
+
+    def _init_check_requests_users(self):
+        # Define users for check_requests default list.
+        self.view_only_lab = USERS[1]
+        self.lab_submitter = USERS[2]
+        # Don't want disabled account to be part of default USERS list.
+        self.disabled_user = 'encoded.test3@gmail.com'
+        self.view_only_admin = USERS[3]
+        # Common response pattern for check_requests, extracted here to avoid
+        # repetition.
+        self._released_expected_response = [('Public', 200),
+                                            (self.view_only_lab, 200),
+                                            (self.lab_submitter, 200),
+                                            (self.disabled_user, 401),
+                                            (self.view_only_admin, 200),
+                                            ('admin', 200)]
+        # For objects owned by lab.
+        self._unreleased_internal_expected_response = [('Public', 403),
+                                                       (self.view_only_lab, 200),
+                                                       (self.lab_submitter, 200),
+                                                       (self.disabled_user, 401),
+                                                       (self.view_only_admin, 200),
+                                                       ('admin', 200)]
+        self._unreleased_other_expected_response = [('Public', 403),
+                                                    (self.view_only_lab, 403),
+                                                    (self.lab_submitter, 403),
+                                                    (self.disabled_user, 401),
+                                                    (self.view_only_admin, 200),
+                                                    ('admin', 200)]
+        self.edit_by_lab = [(self.view_only_lab, 422),
+                            (self.lab_submitter, 200)]
+        self.no_edit_by_lab = [(self.view_only_lab, 422),
+                               (self.lab_submitter, 422)]
+        self.no_edit_non_object_page = [(self.view_only_lab, 403),
+                                        (self.lab_submitter, 403)]
+
+    # Default action lists for QANCODE to check.
+    def _init_default_actions(self):
+        self.compare_facets_default_actions = [
+            '/search/?type=Experiment',
+            '/search/?type=File',
+            '/search/?type=AntibodyLot',
+            '/search/?type=Biosample',
+            '/search/?type=Dataset',
+            '/search/?type=FileSet',
+            '/search/?type=Annotation',
+            '/search/?type=Series',
+            '/search/?type=OrganismDevelopmentSeries',
+            '/search/?type=UcscBrowserComposite',
+            '/search/?type=ReferenceEpigenome',
+            '/search/?type=Project',
+            '/search/?type=ReplicationTimingSeries',
+            '/search/?type=PublicationData',
+            '/search/?type=MatchedSet',
+            '/search/?type=TreatmentConcentrationSeries',
+            '/search/?type=TreatmentTimeSeries',
+            '/search/?type=Target',
+            '/search/?type=Pipeline',
+            '/search/?type=Publication',
+            '/search/?type=Software',
+            '/matrix/?type=Experiment',
+            '/matrix/?type=Annotation'
+        ]
+        self.find_differences_default_actions = [
+            ('/', None),
+            ('/targets/?status=deleted', None),
+            ('/antibodies/?status=deleted', None),
+            ('/search/?type=Biosample&status=deleted', None),
+            ('/experiments/ENCSR000CWD/', None),
+            ('/biosamples/ENCBS574ZRE/', None),
+            ('/biosamples/ENCBS883DWI/', None),
+            ('/experiments/ENCSR985KAT/', None),
+            ('/biosamples/ENCBS298YPF/', None),
+            ('/biosamples/ENCBS142DVU/', None),
+            ('/biosamples/ENCBS562NPI/', None),
+            ('/human-donors/ENCDO999JZG/', None),
+            ('/biosamples/ENCBS615YKY/', None),
+            ('/search/?searchTerm=puf60&type=Target', None),
+            ('/experiments/ENCSR502NRF/', None),
+            ('/experiments/ENCSR000AEH/', None),
+            ('/search/?searchTerm=ENCSR000AEH&type=Experiment', None),
+            ('/experiments/ENCSR000CPG/', None),
+            ('/search/?searchTerm=ENCSR000CPG&type=Experiment', None),
+            ('/experiments/ENCSR000BPF/', None),
+            ('/search/?searchTerm=ENCSR000BPF&type=Experiment', None),
+            ('/experiments/ENCSR178NTX/', None),
+            ('/experiments/ENCSR651NGR/', None),
+            ('/search/?searchTerm=ENCSR651NGR&type=Experiment', None),
+            ('/antibodies/ENCAB000AEH/', None),
+            ('/search/?searchTerm=ENCAB000AEH&type=AntibodyLot', None),
+            ('/antibodies/ENCAB000AIW/', None),
+            ('/search/?searchTerm=ENCAB000AIW&type=AntibodyLot', None),
+            ('/biosamples/ENCBS000AAA/', None),
+            ('/search/?searchTerm=ENCBS000AAA&type=Biosample', None),
+            ('/biosamples/ENCBS030ENC/', None),
+            ('/search/?searchTerm=ENCBS030ENC', None),
+            ('/biosamples/ENCBS098ENC/', None),
+            ('/search/?searchTerm=ENCBS098ENC&type=Biosample', None),
+            ('/biosamples/ENCBS619ENC/', None),
+            ('/search/?searchTerm=ENCBS619ENC&type=Biosample', None),
+            ('/biosamples/ENCBS286AAA/', None),
+            ('/search/?searchTerm=ENCBS286AAA&type=Biosample', None),
+            ('/biosamples/ENCBS314VPT/', None),
+            ('/search/?searchTerm=ENCBS314VPT&type=Biosample', None),
+            ('/biosamples/ENCBS808BUA/', None),
+            ('/search/?searchTerm=ENCBS808BUA&type=Biosample', None),
+            ('/targets/AARS-human/', None),
+            ('/targets/FLAG-GABP-human/', None),
+            ('/search/?type=Target&name=AARS-human', None),
+            ('/ucsc-browser-composites/ENCSR707NXZ/', None),
+            ('/treatment-time-series/ENCSR210PYP/', None),
+            ('/search/?searchTerm=WASP&type=Software', None),
+            ('/publications/67e606ae-abe7-4510-8ebb-cfa1fefe5cfa/', None),
+            ('/search/?searchTerm=PMID%3A25164756', None),
+            ('/biosamples/ENCBS632MTU/', None),
+            ('/biosamples/ENCBS464EKT/', None),
+            ('/annotations/ENCSR790GQB/', None),
+            ('/publications/b2e859e6-3ee7-4274-90be-728e0faaa8b9/', None),
+            ('/pipelines/', None),
+            ('/pipelines/ENCPL210QWH/', None),
+            ('/pipelines/ENCPL002LPE/', None),
+            ('/pipelines/ENCPL002LSE/', None),
+            ('/rna-seq/long-rnas/', None),
+            ('/pipelines/ENCPL337CSA/', None),
+            ('/rna-seq/small-rnas/', None),
+            ('/pipelines/ENCPL444CYA/', None),
+            ('/microrna/microrna-seq/', None),
+            ('/pipelines/ENCPL278BTI/', None),
+            ('/microrna/microrna-counts/', None),
+            ('/pipelines/ENCPL122WIM/', None),
+            ('/rampage/', None),
+            ('/pipelines/ENCPL220NBH/', None),
+            ('/pipelines/ENCPL272XAE/', None),
+            ('/pipelines/ENCPL272XAE/', None),
+            ('/chip-seq/histone/', None),
+            ('/pipelines/ENCPL138KID/', None),
+            ('/pipelines/ENCPL493SGC/', None),
+            ('/chip-seq/transcription_factor/', None),
+            ('/pipelines/ENCPL001DNS/', None),
+            ('/pipelines/ENCPL002DNS/', None),
+            ('/data-standards/dnase-seq/', None),
+            ('/atac-seq/', None),
+            ('/pipelines/ENCPL985BLO/', None),
+            ('/data/annotations/', None),
+            ('/help/rest-api/', None),
+            ('/about/experiment-guidelines/', None),
+            ('/data-standards/terms/', None)
+        ]
+        self.check_trackhubs_default_actions = [
+            ('/experiments/ENCSR502NRF/',
+             OpenUCSCGenomeBrowserGRCh38),
+            ('/experiments/ENCSR502NRF/',
+             OpenUCSCGenomeBrowserHG19),
+            ('/experiments/ENCSR985KAT/',
+             OpenUCSCGenomeBrowserHG19),
+            ('/experiments/ENCSR426UUG/',
+             OpenUCSCGenomeBrowserGRCh38),
+            ('/experiments/ENCSR293WTN/',
+             OpenUCSCGenomeBrowserMM9),
+            ('/experiments/ENCSR335LKF/',
+             OpenUCSCGenomeBrowserMM10),
+            ('/experiments/ENCSR922ESH/',
+             OpenUCSCGenomeBrowserDM3),
+            ('/experiments/ENCSR671XAK/',
+             OpenUCSCGenomeBrowserDM6),
+            ('/experiments/ENCSR422XRE/',
+             OpenUCSCGenomeBrowserCE10),
+            ('/experiments/ENCSR686FKU/',
+             OpenUCSCGenomeBrowserCE11),
+            ('/publication-data/ENCSR764APB/',
+             OpenUCSCGenomeBrowserHG19),
+            ('/projects/ENCSR295OIE/',
+             OpenUCSCGenomeBrowserHG19),
+            ('/annotations/ENCSR212BHV/',
+             OpenUCSCGenomeBrowserHG19),
+            ('/experiments/ENCSR000CJR/',
+             OpenUCSCGenomeBrowserHG19),
+            ('/search/?type=Experiment&assembly=hg19&target.investigated_as=RNA+binding+protein&assay_title=ChIP-seq&replicates.library.biosample.biosample_type=primary+cell',
+             OpenUCSCGenomeBrowserHG19),
+            ('/search/?type=Experiment&assembly=GRCh38&assay_title=shRNA+RNA-seq&target.investigated_as=transcription+factor&month_released=October%2C+2014',
+             OpenUCSCGenomeBrowserGRCh38),
+            ('/search/?type=Experiment&assembly=mm9&assay_title=Repli-chip',
+             OpenUCSCGenomeBrowserMM9),
+            ('/search/?type=Experiment&assembly=mm10&assay_title=microRNA-seq&month_released=January%2C+2016',
+             OpenUCSCGenomeBrowserMM10),
+            ('/search/?type=Experiment&assembly=dm3&status=released&replicates.library.biosample.biosample_type=whole+organisms&assay_title=total+RNA-seq',
+             OpenUCSCGenomeBrowserDM3),
+            ('/search/?type=Experiment&assembly=dm6&replicates.library.biosample.life_stage=wandering+third+instar+larva',
+             OpenUCSCGenomeBrowserDM6),
+            ('/search/?type=Experiment&assembly=ce10&target.investigated_as=transcription+factor&replicates.library.biosample.life_stage=L4+larva',
+             OpenUCSCGenomeBrowserCE10),
+            ('/search/?type=Experiment&assembly=ce11&target.investigated_as=recombinant+protein&replicates.library.biosample.life_stage=late+embryonic&replicates.library.biosample.life_stage=L4+larva',
+             OpenUCSCGenomeBrowserCE11),
+            ('/search/?searchTerm=hippocampus&type=Experiment',
+             OpenUCSCGenomeBrowserHG19)
+        ]
+        self.check_permissions_default_actions = [
+            ('/experiments/ENCSR524OCB/', None),
+            ('/experiments/ENCSR000EFT/', None),
+            ('/biosamples/ENCBS643IYW/', None),
+            ('/experiments/ENCSR466YGC/', None),
+            ('/experiments/ENCSR255XZG/', None),
+            ('/experiments/ENCSR115BCB/', None),
+            ('/files/ENCFF752JWY/', None),
+            ('/targets/2L52.1-celegans/', None),
+            ('/targets/CG15455-dmelanogaster/', None),
+            ('/software/dnase-eval-bam-se/', None),
+            ('/software/atac-seq-software-tools/', None),
+            ('/software/trimAdapters.py/', None),
+            ('/software/bigwigaverageoverbed/', None),
+            ('/pipelines/ENCPL493SGC/', None),
+            ('/pipelines/ENCPL035XIO/', None),
+            ('/pipelines/ENCPL568PWV/', None),
+            ('/pipelines/e02448b1-9706-4e7c-b31b-78c921d58f0b/', None),
+            ('/pipelines/ENCPL734EDH/', None),
+            ('/pipelines/ENCPL983UFZ/', None),
+            ('/pipelines/ENCPL631XPY/', None),
+            ('/publications/b2e859e6-3ee7-4274-90be-728e0faaa8b9/', None),
+            ('/publications/a4db2c6d-d1a3-4e31-b37b-5cc7d6277548/', None),
+            ('/publications/16c77add-1bfb-424b-8cab-498ac1e5f6ed/', None),
+            ('/publications/da2f7542-3d99-48f6-a95d-9907dd5e2f81/', None),
+            ('/internal-data-use-policy/', None),
+            ('/tutorials/encode-users-meeting-2016/logistics/', None),
+            ('/2017-06-09-release/', None)
+        ]
+        self.check_downloads_default_actions = [
+            ('/experiments/ENCSR810WXH/', DownloadBEDFileFromTable),
+            ('/experiments/ENCSR966YYJ/',
+             DownloadBEDFileFromModal),
+            ('/experiments/ENCSR810WXH/',
+             DownloadGraphFromExperimentPage),
+            ('/experiments/ENCSR810WXH/',
+             DownloadDocuments),
+            ('/ucsc-browser-composites/ENCSR707NXZ/',
+             DownloadDocuments),
+            ('/antibodies/ENCAB749XQY/',
+             DownloadDocumentsFromAntibodyPage),
+            ('/report/?searchTerm=nose&type=Biosample',
+             DownloadTSVFromReportPage),
+            ('/search/?type=Experiment&searchTerm=nose',
+             DownloadMetaDataFromSearchPage),
+            ('/files/ENCFF931OLL/',
+             DownloadFileFromFilePage),
+            ('/files/ENCFF291ELS/', DownloadFileFromFilePage)
+        ]
+        # List of tuples:
+        # ('suburl', [{payloads}], [(user, expected_status_code)])
+        self.check_requests_default_actions = {
+            'patch': [
+                ('/experiments/ENCSR000CUS/',
+                 [{'description': 'test'}],
+                 [('Public', 400),
+                  (self.view_only_admin, 403)]),
+                ('/experiments/ENCSR000CUS/',
+                 [{'status': 'deleted'},
+                  {'status': 'archived'},
+                  {'status': 'proposed'},
+                  {'status': 'ready for review'},
+                  {'status': 'released'},
+                  {'status': 'started'},
+                  {'status': 'submitted'},
+                  {'status': 'replaced'}],
+                 [('admin', 200)]),
+                ('/experiments/ENCSR035DLJ/',
+                 [{'alternate_accessions': ['ENCSR000CUS']},
+                  {'alternate_accessions': []}],
+                 [('admin', 200)]),
+                ('35f91f16-dcef-4ab2-90bd-3928b0db9a60',
+                 [{'status': 'revoked'}],
+                 [('admin', 200)]),
+                ('/files/ENCFF002BYE/',
+                 [{'status': 'deleted'},
+                  {'status': 'in progress'},
+                  {'status': 'released'},
+                  {'status': 'replaced'}],
+                 [('admin', 200)]),
+                ('4dc1fbd3-6692-42fa-b710-03eaba9263c1',
+                 [{'status': 'revoked'}],
+                 [('admin', 200)]),
+                ('/experiments/ENCSR524OCB/',
+                 [{'status': 'submitted'}],
+                 self.no_edit_by_lab),
+                ('/experiments/ENCSR000EFT/',
+                 [{'status': 'revoked'}],
+                 self.no_edit_by_lab),
+                ('/biosamples/ENCBS643IYW/',
+                 [{'status': 'in progress'}],
+                 self.edit_by_lab),
+                ('/libraries/ENCLB220SGE/',
+                 [{'status': 'in progress'}],
+                 self.no_edit_by_lab),
+                ('/experiments/ENCSR466YGC/',
+                 [{'status': 'in progress'}],
+                 self.no_edit_by_lab),
+                ('/experiments/ENCSR255XZG/',
+                 [{'status': 'in progress'}],
+                 self.no_edit_by_lab),
+                ('/experiments/ENCSR115BCB/',
+                 [{'status': 'in progress'}],
+                 self.no_edit_by_lab),
+                ('/files/ENCFF851EYG/',
+                 [{'status': 'in progress'}],
+                 self.edit_by_lab),
+                ('/targets/2L52.1-celegans/',
+                 [{'status': 'in progress'}],
+                 self.no_edit_by_lab),
+                ('/software/dnase-eval-bam-se/',
+                 [{'status': 'in progress'}],
+                 self.no_edit_by_lab),
+                ('/software/atac-seq-software-tools/',
+                 [{'status': 'in progress'}],
+                 self.no_edit_by_lab),
+                ('/software/trimAdapters.py/',
+                 [{'status': 'in progress'}],
+                 self.no_edit_by_lab),
+                ('/software/bigwigaverageoverbed/',
+                 [{'status': 'in progress'}],
+                 self.no_edit_by_lab),
+                ('/pipelines/ENCPL493SGC/',
+                 [{'status': 'in progress'}],
+                 self.no_edit_by_lab),
+                ('/pipelines/ENCPL035XIO/',
+                 [{'status': 'in progress'}],
+                 self.no_edit_by_lab),
+                ('/pipelines/ENCPL568PWV/',
+                 [{'status': 'in progress'}],
+                 self.no_edit_by_lab),
+                ('/pipelines/e02448b1-9706-4e7c-b31b-78c921d58f0b/',
+                 [{'status': 'in progress'}],
+                 self.no_edit_by_lab),
+                ('/pipelines/ENCPL855QIE/',
+                 [{'status': 'in progress'}],
+                 self.no_edit_by_lab),
+                ('/pipelines/ENCPL983UFZ/',
+                 [{'status': 'in progress'}],
+                 self.no_edit_by_lab),
+                ('/pipelines/ENCPL631XPY/',
+                 [{'status': 'in progress'}],
+                 self.no_edit_by_lab),
+                ('/publications/b2e859e6-3ee7-4274-90be-728e0faaa8b9/',
+                 [{'status': 'in progress'}],
+                 self.no_edit_by_lab),
+                ('/publications/a4db2c6d-d1a3-4e31-b37b-5cc7d6277548/',
+                 [{'status': 'in progress'}],
+                 self.no_edit_by_lab),
+                ('/publications/16c77add-1bfb-424b-8cab-498ac1e5f6ed/',
+                 [{'status': 'in progress'}],
+                 self.no_edit_by_lab),
+                ('/publications/da2f7542-3d99-48f6-a95d-9907dd5e2f81/',
+                 [{'status': 'in progress'}],
+                 self.no_edit_by_lab),
+                ('/internal-data-use-policy/',
+                 [{'status': 'in progress'}],
+                 self.no_edit_non_object_page),
+                ('/tutorials/encode-users-meeting-2016/logistics/',
+                 [{'status': 'in progress'}],
+                 self.no_edit_non_object_page),
+                ('/2017-06-09-release/',
+                 [{'status': 'in progress'}],
+                 self.no_edit_by_lab)
+            ],
+            'post': [
+                ('/experiments/',
+                 [{'description': 'test post experiment',
+                   'assay_term_name': 'ChIP-seq',
+                   'biosample_term_id': 'CL:0010001',
+                   'biosample_type': 'primary cell',
+                   'biosample_term_name': 'Stromal cell of bone marrow',
+                   'target': '/targets/SMAD6-human/',
+                   'award': '/awards/U41HG006992/',
+                   'lab': '/labs/thomas-gingeras/',
+                   'references': ['PMID:18229687', 'PMID:25677182']}],
+                 [('Public', 400),
+                  (self.lab_submitter, 201),
+                  ('admin', 201)])
+            ],
+            'get': [
+                ('/experiments/ENCSR082IHY/',
+                 [None],
+                 [('Public', 403),
+                  (self.lab_submitter, 403),
+                  ('admin', 200)]),
+                ('/experiments/ENCSR000CUS',
+                 [None],
+                 [('Public', 200),
+                  (self.lab_submitter, 200),
+                  ('admin', 200)]),
+                ('/experiments/ENCSR524OCB/',
+                 [None],
+                 self._unreleased_internal_expected_response),
+                ('/experiments/ENCSR000EFT/',
+                 [None],
+                 self._released_expected_response),
+                ('/biosamples/ENCBS643IYW/',
+                 [None],
+                 self._unreleased_internal_expected_response),
+                ('/libraries/ENCLB220SGE/',
+                 [None],
+                 self._unreleased_internal_expected_response),
+                ('/experiments/ENCSR466YGC/',
+                 [None],
+                 self._unreleased_internal_expected_response),
+                ('/experiments/ENCSR255XZG/',
+                 [None],
+                 self._unreleased_other_expected_response),
+                ('/experiments/ENCSR115BCB/',
+                 [None],
+                 self._unreleased_other_expected_response),
+                ('/files/ENCFF851EYG/',
+                 [None],
+                 self._unreleased_internal_expected_response),
+                ('/targets/2L52.1-celegans/',
+                 [None],
+                 self._released_expected_response),
+                ('/software/dnase-eval-bam-se/',
+                 [None],
+                 self._unreleased_other_expected_response),
+                ('/software/atac-seq-software-tools/',
+                 [None],
+                 self._unreleased_other_expected_response),
+                ('/software/trimAdapters.py/',
+                 [None],
+                 self._unreleased_internal_expected_response),
+                ('/software/bigwigaverageoverbed/',
+                 [None],
+                 self._released_expected_response),
+                ('/pipelines/ENCPL493SGC/',
+                 [None],
+                 self._released_expected_response),
+                ('/pipelines/ENCPL035XIO/',
+                 [None],
+                 self._released_expected_response),
+                ('/pipelines/ENCPL568PWV/',
+                 [None],
+                 self._unreleased_other_expected_response),
+                ('/pipelines/e02448b1-9706-4e7c-b31b-78c921d58f0b/',
+                 [None],
+                 self._unreleased_other_expected_response),
+                ('/pipelines/ENCPL855QIE/',
+                 [None],
+                 self._unreleased_internal_expected_response),
+                ('/pipelines/ENCPL983UFZ/',
+                 [None],
+                 self._unreleased_other_expected_response),
+                ('/pipelines/ENCPL631XPY/',
+                 [None],
+                 self._released_expected_response),
+                ('/publications/b2e859e6-3ee7-4274-90be-728e0faaa8b9/',
+                 [None],
+                 self._released_expected_response),
+                ('/publications/a4db2c6d-d1a3-4e31-b37b-5cc7d6277548/',
+                 [None],
+                 self._unreleased_internal_expected_response),
+                ('/publications/16c77add-1bfb-424b-8cab-498ac1e5f6ed/',
+                 [None],
+                 self._unreleased_internal_expected_response),
+                ('/publications/da2f7542-3d99-48f6-a95d-9907dd5e2f81/',
+                 [None],
+                 self._unreleased_other_expected_response),
+                ('/internal-data-use-policy/',
+                 [None],
+                 self._unreleased_internal_expected_response),
+                ('/tutorials/encode-users-meeting-2016/logistics/',
+                 [None],
+                 self._unreleased_other_expected_response),
+                ('/2017-06-09-release/',
+                 [None],
+                 self._released_expected_response)
+            ]
+        }
+        self.check_tools_default_actions = [
+            {'name': 'ENCODE_get_fields.py',
+             'command': '{} {} {}  --infile ENCSR000CUS --field status',
+             'expected_output': 'accession\tstatus\r\nENCSR000CUS\trevoked\r\n'},
+            {'name': 'ENCODE_patch_set.py',
+             'command': '{} {} {} --accession ENCSR000CUS --field status --data revoked',
+             'expected_output': 'OBJECT: ENCSR000CUS\nOLD DATA: status'
+             ' revoked\nNEW DATA: status revoked'},
+            {'name': 'ENCODE_release.py',
+             'command': '{} {} {} --infile ENCSR000CUS',
+             'expected_output': 'Data written to file Release_report.txt'},
+            {'name': 'ENCODE_submit_files.py',
+             'command': '{} {} permissions_qa_scripts/Test_submit_files.csv {}',
+             'expected_output': "'file_size': 23972104"}
+        ]
